requirements.txt
Flask==1.1.1
Werkzeug==0.16.1
google-cloud-bigquery==1.24
numpy==1.17
